reorganize example
