+++ examples/usa_evanston_yan/nexus_em_ebsd_patterns_metadata.yaml
